RVV: refactor: seperate vcompute def and impl
