+++ src/isa/riscv64/instr/rvv/vcompute.h
@@ -1,276 +1,4 @@
-#include "cpu/exec.h"
-#include "vreg.h"
-#include "../local-include/csr.h"
-#include <stdio.h>
-#include "../local-include/intr.h"
-#include "../local-include/rtl.h"
-#include <setjmp.h>
-
-#define id_src (&s->src1)
-#define id_src2 (&s->src2)
-#define id_dest (&s->dest)
-
-#undef s0
-#undef s1
-//#undef s2
-//#undef t0
-
-#define s0    (&tmp_reg[0])
-#define s1    (&tmp_reg[1])
-//#define s2    (&tmp_reg[2])
-//#define t0    (&tmp_reg[3])
-
-enum op_t {
-  ADD, SUB, RSUB, MINU, MIN, MAXU, MAX, AND,
-  OR, XOR, RGATHER, SLIDEUP, SLIDEDOWN, ADC, MADC, SBC,
-  MSBC, MERGE, MSEQ, MSNE, MSLTU, MSLT, MSLEU, MSLE,
-  MSGTU, MSGT, SADDU, SADD, SSUBU, SSUB, AADD, SLL,
-  ASUB, SMUL, SRL, SRA, SSRL, SSRA, NSRL, NSRA,
-  NCLIPU, NCLIP, WREDSUMU, WREDSUM, DOTU, DOT, WSMACCU, WSMACC,
-  WSMAXXSU, WSMACCUS,
-
-  REDSUM, REDAND, REDOR, REDXOR, REDMINU, REDMIN, REDMAXU, REDMAX,
-  EXT_X_V, MV_S_X, SLIDE1UP, SLIDE1DOWN, MPOPC, VMFIRST, MUNARYO, COMPRESS,
-  MANDNOT, MAND, MOR, MXOR, MORNOT, MNAND, MNOR, MXNOR,
-  DIVU, DIV, REMU, REM, MULHU, MUL, MULHSU, MULH,
-  MADD, NMSUB, MACC, NMSAC, WADDU, WADD, WSUBU, WSUB,
-  WADDU_W, WADD_W, WSUBU_W, WSUB_W, WMULU, WMULSU, WMUL, WMACCU,
-  WNMACC, WMACCSU, WMACCUS,
-};
-
-#define ARTHI(opcode, is_signed) arthimetic_instr(opcode, is_signed, 0, s);
-#define ARTHI_COMP(opcode, is_signed) arthimetic_instr(opcode, is_signed, 1, s);
-void arthimetic_instr(int opcode, int is_signed, int dest_reg, Decode *s) {
-  vp_set_dirty();
-  int idx;
-  for(idx = vstart->val; idx < vl->val; idx ++) {
-    // mask
-    rtlreg_t mask = get_mask(0, idx, vtype->vsew, vtype->vlmul);
-    if(s->vm == 0) {
-      // merge instr will exec no matter mask or not
-      // masked and mask off exec will left dest unmodified.
-      if(opcode != MERGE && mask==0) continue;
-    } else if(opcode == MERGE) {
-      mask = 1; // merge(mv) get the first operand (s1, rs1, imm);
-    }
-
-    // operand - vs2
-    get_vreg(id_src2->reg, idx, s0, vtype->vsew, vtype->vlmul, is_signed, 1);
-     if(is_signed) rtl_sext(s, s0, s0, 1 << vtype->vsew);
-
-    // operand - s1 / rs1 / imm
-    switch (s->src_vmode) {
-      case SRC_VV : 
-        get_vreg(id_src->reg, idx, s1, vtype->vsew, vtype->vlmul, is_signed, 1);
-        if(is_signed) rtl_sext(s, s1, s1, 1 << vtype->vsew);
-        break;
-      case SRC_VS :   
-        rtl_lr(s, &(id_src->val), id_src1->reg, 4);
-        rtl_mv(s, s1, &id_src->val); 
-        if(is_signed) rtl_sext(s, s1, s1, 1 << vtype->vsew);
-        break;
-      case SRC_VI :
-        if(is_signed) rtl_li(s, s1, s->isa.instr.v_opv2.v_simm5);
-        else          rtl_li(s, s1, s->isa.instr.v_opv3.v_imm5 );
-        break;
-    }
-
-    // op
-    switch (opcode) {
-      case ADD : rtl_add(s, s1, s0, s1); break;
-      case SUB : rtl_sub(s, s1, s0, s1); break;
-      case RSUB: rtl_sub(s, s1, s1, s0); break;
-      case AND : rtl_and(s, s1, s0, s1); break;
-      case OR  : rtl_or(s, s1, s0, s1); break;
-      case XOR : rtl_xor(s, s1, s0, s1); break;
-      case SLL :
-        rtl_andi(s, s1, s1, s->v_width*8-1); //low lg2(SEW) is valid
-        //rtl_sext(s0, s0, 8 - (1 << vtype->vsew)); //sext first
-        rtl_shl(s, s1, s0, s1); break;
-      case SRL :
-        rtl_andi(s, s1, s1, s->v_width*8-1); //low lg2(SEW)
-        rtl_shr(s, s1, s0, s1); break;
-      case SRA :
-        rtl_andi(s, s1, s1, s->v_width*8-1); //low lg2(SEW)
-        rtl_sext(s, s0, s0, s->v_width);
-        rtl_sar(s, s1, s0, s1); break;
-      case MULHU : 
-        *s1 = (uint64_t)(((__uint128_t)(*s0) * (__uint128_t)(*s1))>>(s->v_width*8));
-        break;
-      case MUL : rtl_mulu_lo(s, s1, s0, s1); break;
-      case MULHSU :
-        rtl_sext(s, t0, s0, s->v_width);
-        rtl_sari(s, t0, t0, s->v_width*8-1);
-        rtl_and(s, t0, s1, t0);
-        *s1 = (uint64_t)(((__uint128_t)(*s0) * (__uint128_t)(*s1))>>(s->v_width*8));
-        rtl_sub(s, s1, s1, t0);
-        break;
-      case MULH :
-        rtl_sext(s, s0, s0, s->v_width);
-        rtl_sext(s, s1, s1, s->v_width);
-        *s1 = (uint64_t)(((__int128_t)(sword_t)(*s0) * (__int128_t)(sword_t)(*s1))>>(s->v_width*8));
-        break;
-      case MACC : 
-        rtl_mulu_lo(s, s1, s0, s1);
-        get_vreg(id_dest->reg, idx, s0, vtype->vsew, vtype->vlmul, is_signed, 1);
-        rtl_add(s, s1, s1, s0);
-        break;
-      case NMSAC :
-        rtl_mulu_lo(s, s1, s0, s1);
-        get_vreg(id_dest->reg, idx, s0, vtype->vsew, vtype->vlmul, is_signed, 1);
-        rtl_sub(s, s1, s0, s1);
-        break;
-      case MADD :
-        get_vreg(id_dest->reg, idx, s0, vtype->vsew, vtype->vlmul, is_signed, 1);
-        rtl_mulu_lo(s, s1, s0, s1);
-        get_vreg(id_src2->reg, idx, s0, vtype->vsew, vtype->vlmul, is_signed, 1);
-        rtl_add(s, s1, s1, s0);
-        break;
-      case NMSUB :
-        get_vreg(id_dest->reg, idx, s0, vtype->vsew, vtype->vlmul, is_signed, 1);
-        rtl_mulu_lo(s, s1, s1, s0);
-        get_vreg(id_src2->reg, idx, s0, vtype->vsew, vtype->vlmul, is_signed, 1);
-        rtl_sub(s, s1, s0, s1);
-        break;
-      case DIVU :
-        if(*s1 == 0) rtl_li(s, s1, ~0lu);
-        else rtl_divu_q(s, s1, s0, s1);
-        break;
-      case DIV :
-        rtl_sext(s, s0, s0, s->v_width);
-        rtl_sext(s, s1, s1, s->v_width);
-        if(*s1 == 0) rtl_li(s, s1, ~0lu);
-        else if(*s0 == 0x8000000000000000LL && *s1 == -1) //may be error
-          rtl_mv(s, s1, s0);
-        else rtl_divs_q(s, s1, s0, s1);
-        break;
-      case REMU :
-        if (*s1 == 0) rtl_mv(s, s1, s0);
-        else rtl_divu_r(s, s1, s0, s1);
-        break;
-      case REM :
-        rtl_sext(s, s0, s0, s->v_width);
-        rtl_sext(s, s1, s1, s->v_width);
-        if(*s1 == 0) rtl_mv(s, s1, s0);
-        else if(*s1 == 0x8000000000000000LL && *s1 == -1) //may be error
-          rtl_li(s, s1, 0);
-        else rtl_divs_r(s, s1, s0, s1);
-        break;
-      case MERGE : rtl_mux(s, s1, &mask, s1, s0); break;
-      case MSEQ  : rtl_setrelop(s, RELOP_EQ,  s1, s0, s1); break;
-      case MSNE  : rtl_setrelop(s, RELOP_NE,  s1, s0, s1); break;
-      case MSLTU : rtl_setrelop(s, RELOP_LTU, s1, s0, s1); break;
-      case MSLT  : rtl_setrelop(s, RELOP_LT,  s1, s0, s1); break;
-      case MSLEU : rtl_setrelop(s, RELOP_LEU, s1, s0, s1); break;
-      case MSLE  : rtl_setrelop(s, RELOP_LE,  s1, s0, s1); break;
-      case MSGTU : rtl_setrelop(s, RELOP_GTU, s1, s0, s1); break;
-      case MSGT  : rtl_setrelop(s, RELOP_GT,  s1, s0, s1); break;
-    }
-
-    // store to vrf
-    if(dest_reg == 1) 
-      set_mask(id_dest->reg, idx, *s1, vtype->vsew, vtype->vlmul);
-    else 
-      set_vreg(id_dest->reg, idx, *s1, vtype->vsew, vtype->vlmul, 1);
-  }
-
-  // idx gt the vl need to be zeroed.
-  // int vlmax = ((VLEN >> 3) >> vtype->vsew) << vtype->vlmul;
-  // for(idx = vl->val; idx < vlmax; idx ++) {
-  //   rtl_li(s1, 0);
-  //   if(dest_reg == 1) 
-  //     set_mask(id_dest->reg, idx, s1, vtype->vsew, vtype->vlmul);
-  //   else 
-  //     set_vreg(id_dest->reg, idx, s1, vtype->vsew, vtype->vlmul, 1);
-  // }
-
-  // TODO: the idx larger than vl need reset to zero.
-  rtl_li(s, s0, 0);
-  vcsr_write(IDXVSTART, s0);
-}
-
-#define MASKINSTR(opcode) mask_instr(opcode, s);
-void mask_instr(int opcode, Decode *s) {
-  vp_set_dirty();
-  int idx;
-  for(idx = vstart->val; idx < vl->val; idx++) {
-    // operand - vs2
-    *s0 = get_mask(id_src2->reg, idx, vtype->vsew, vtype->vlmul); // unproper usage of s0
-    *s0 &= 1; // only LSB
-
-    // operand - s1
-    *s1 = get_mask(id_src->reg, idx, vtype->vsew, vtype->vlmul); // unproper usage of s1
-    *s1 &= 1; // only LSB
-
-    // op
-    switch (opcode) {
-      case MAND    : rtl_and(s, s1, s0, s1); break;
-      case MNAND   : rtl_and(s, s1, s0, s1);
-                     *s1 = !(*s1); break;
-      case MANDNOT : *s1 = !(*s1); // unproper usage of not
-                     rtl_and(s, s1, s0, s1); break;
-      case MXOR    : rtl_xor(s, s1, s0, s1); break;
-      case MOR     : rtl_or(s, s1, s0, s1); break;
-      case MNOR    : rtl_or(s, s1, s0, s1);
-                     *s1 = !(*s1); break;
-      case MORNOT  : *s1 = !(*s1);
-                     rtl_or(s, s1, s0, s1); break;
-      case MXNOR   : rtl_xor(s, s1, s0, s1);
-                     *s1 = !(*s1); break;
-      default      : longjmp_raise_intr(EX_II);
-    }
-    // store to vrf
-    *s1 &= 1; // make sure the LSB
-    set_mask(id_dest->reg, idx, *s1, vtype->vsew, vtype->vlmul);
-  }
-
-  int vlmax = ((VLEN >> 3) >> vtype->vsew) << vtype->vlmul;
-  rtl_li(s, s1, 0);
-  for( idx = vl->val; idx < vlmax; idx++) {  
-    set_mask(id_dest->reg, idx, *s1, vtype->vsew, vtype->vlmul);
-  }
-  vcsr_write(IDXVSTART, s1);
-}
-
-#define REDInstr(opcode, is_signed) reduction_instr(opcode, is_signed, s);
-void reduction_instr(int opcode, int is_signed, Decode *s) {
-  vp_set_dirty();
-  get_vreg(id_src->reg, 0, s1, vtype->vsew, vtype->vlmul, is_signed, 0);
-  if(is_signed) rtl_sext(s, s1, s1, 1 << vtype->vsew);
-
-  int idx;
-  for(idx = vstart->val; idx < vl->val; idx ++) {
-    rtlreg_t mask = get_mask(0, idx, vtype->vsew, vtype->vlmul);
-    if(s->vm == 0) {
-      // merge instr will exec no matter mask or not
-      // masked and mask off exec will left dest unmodified.
-      if(opcode != MERGE && mask==0) continue;
-    } else if(opcode == MERGE) {
-      mask = 1; // merge(mv) get the first operand (s1, rs1, imm);
-    }
-    // operand - vs2
-    get_vreg(id_src2->reg, idx, s0, vtype->vsew, vtype->vlmul, is_signed, 1);
-    if(is_signed) rtl_sext(s, s0, s0, 1 << vtype->vsew);
-
-
-    // op
-    switch (opcode) {
-      case REDSUM : rtl_add(s, s1, s0, s1); break;
-      case REDOR  : rtl_or(s, s1, s0, s1); break;
-      case REDAND : rtl_and(s, s1, s0, s1); break;
-      case REDXOR : rtl_xor(s, s1, s0, s1); break;
-      //  case MIN : 
-      // MINU is hard to achieve parallel
-    }
-
-  }
-  set_vreg(id_dest->reg, 0, *s1, vtype->vsew, vtype->vlmul, 0);
-  
-  int vlmax =  ((VLEN >> 3) >> vtype->vsew);
-  for(int i=1; i<vlmax; i++) {
-    set_vreg(id_dest->reg, i, 0, vtype->vsew, vtype->vlmul, 0);
-  }
-}
+#include "vcompute_impl.h"
 
 // #define PERMInstr(opcode) permutation_instr(opcode, pc);
 // static void permutation_instr(int opcode, vaddr_t* pc) {
@@ -797,8 +525,4 @@ def_EHelper(vwmaccus) {
   longjmp_raise_intr(EX_II);
 }
 
-// dirty job here
-#undef s0
-#undef s1
-#define s0 &ls0
-#define s1 &ls1
+
